+++ test/bench/granular/code.bench.ts
@@ -2,10 +2,9 @@ import { Suite } from 'bson-bench';
 import {
   getTestDocs,
   runSuiteAndWriteResults,
-  BSON_VERSIONS,
-  BSONEXT_VERSIONS,
   OPERATIONS,
   ITERATIONS,
+  LIBRARY_SPEC,
   WARMUP
 } from './common';
 
@@ -15,18 +14,16 @@ async function main() {
     await getTestDocs('code-with-scope')
   );
 
-  for (const library of BSON_VERSIONS.concat(BSONEXT_VERSIONS)) {
-    for (const operation of OPERATIONS) {
-      for (const documentPath of testDocs) {
-        suite.task({
-          documentPath,
-          library,
-          iterations: ITERATIONS,
-          warmup: WARMUP,
-          operation,
-          options: {}
-        });
-      }
+  for (const operation of OPERATIONS) {
+    for (const documentPath of testDocs) {
+      suite.task({
+        documentPath,
+        library: LIBRARY_SPEC,
+        iterations: ITERATIONS,
+        warmup: WARMUP,
+        operation,
+        options: {}
+      });
     }
   }
   await runSuiteAndWriteResults(suite);
