+++ test/bench/granular/common.ts
@@ -2,13 +2,14 @@ import { Suite } from 'bson-bench';
 import * as path from 'path';
 import * as fs from 'fs/promises';
 
-export const DOCUMENT_ROOT = `${__dirname}/../documents`;
+export const DOCUMENT_ROOT = `${__dirname}/../../documents`;
 export const BSON_VERSIONS = [`bson@6`, `bson@5`, 'bson@4'];
 export const BSONEXT_VERSIONS = ['bson-ext@4.0.0'];
 export const OPERATIONS: ('serialize' | 'deserialize')[] = ['serialize', 'deserialize'];
 export const BOOL = [true, false];
-export const ITERATIONS = 10_000;
-export const WARMUP = 1000;
+
+export const LIBRARY = path.resolve(`${__dirname}/../../../../.`);
+export const LIBRARY_SPEC = `bson:${LIBRARY}`;
 
 export const isDeserialize = (s: string) => s === 'deserialize';
 export async function getTestDocs(type: string) {
@@ -33,3 +34,16 @@ export async function runSuiteAndWriteResults(suite: Suite) {
   await suite.run();
   await suite.writeResults(`${suite.name.toLowerCase()}Results.json`);
 }
+
+export function readEnvVars(): { warmup: number; iterations: number } {
+  const envWarmup = Number(process.env.WARMUP);
+  const envIterations = Number(process.env.ITERATIONS);
+  return {
+    warmup: Number.isSafeInteger(envWarmup) && envWarmup > 0 ? envWarmup : 100_000,
+    iterations: Number.isSafeInteger(envIterations) && envIterations > 0 ? envIterations : 10_000
+  };
+}
+
+const envVars = readEnvVars();
+export const ITERATIONS = envVars.iterations;
+export const WARMUP = envVars.warmup;
