+++ test/bench/granular/date.bench.ts
@@ -1,0 +1,29 @@
+import { Suite } from 'bson-bench';
+import {
+  getTestDocs,
+  runSuiteAndWriteResults,
+  OPERATIONS,
+  ITERATIONS,
+  LIBRARY_SPEC,
+  WARMUP
+} from './common';
+
+async function main() {
+  const suite = new Suite('Date');
+  const testDocs = await getTestDocs('date');
+
+  for (const operation of OPERATIONS) {
+    for (const documentPath of testDocs) {
+      suite.task({
+        documentPath,
+        library: LIBRARY_SPEC,
+        iterations: ITERATIONS,
+        warmup: WARMUP,
+        operation,
+        options: {}
+      });
+    }
+  }
+  await runSuiteAndWriteResults(suite);
+}
+main();
