+++ test/bench/granular/decimal128.bench.ts
@@ -2,10 +2,9 @@ import { Suite } from 'bson-bench';
 import {
   getTestDocs,
   runSuiteAndWriteResults,
-  BSON_VERSIONS,
-  BSONEXT_VERSIONS,
   OPERATIONS,
   ITERATIONS,
+  LIBRARY_SPEC,
   WARMUP
 } from './common';
 
@@ -22,18 +21,16 @@ const OPTIONS = {
 async function main() {
   const suite = new Suite('Decimal128');
   const testDocs = await getTestDocs('decimal128');
-  for (const library of BSON_VERSIONS.concat(BSONEXT_VERSIONS)) {
-    for (const operation of OPERATIONS) {
-      for (const documentPath of testDocs) {
-        suite.task({
-          documentPath,
-          library,
-          iterations: ITERATIONS,
-          warmup: WARMUP,
-          operation,
-          options: OPTIONS[operation]
-        });
-      }
+  for (const operation of OPERATIONS) {
+    for (const documentPath of testDocs) {
+      suite.task({
+        documentPath,
+        library: LIBRARY_SPEC,
+        iterations: ITERATIONS,
+        warmup: WARMUP,
+        operation,
+        options: OPTIONS[operation]
+      });
     }
   }
   await runSuiteAndWriteResults(suite);
