+++ test/bench/granular/int32.bench.ts
@@ -3,8 +3,7 @@ import {
   getTestDocs,
   isDeserialize,
   runSuiteAndWriteResults,
-  BSON_VERSIONS,
-  BSONEXT_VERSIONS,
+  LIBRARY_SPEC,
   OPERATIONS,
   BOOL,
   ITERATIONS,
@@ -25,21 +24,19 @@ const OPTIONS = {
 async function main() {
   const suite = new Suite('Int32');
   const testDocs = await getTestDocs('int32');
-  for (const library of BSON_VERSIONS.concat(BSONEXT_VERSIONS)) {
-    for (const operation of OPERATIONS) {
-      for (const documentPath of testDocs) {
-        for (const promoteValues of BOOL) {
-          suite.task({
-            documentPath,
-            library,
-            iterations: ITERATIONS,
-            warmup: WARMUP,
-            operation,
-            options: isDeserialize(operation)
-              ? { ...OPTIONS[operation], promoteValues }
-              : OPTIONS[operation]
-          });
-        }
+  for (const operation of OPERATIONS) {
+    for (const documentPath of testDocs) {
+      for (const promoteValues of BOOL) {
+        suite.task({
+          documentPath,
+          library: LIBRARY_SPEC,
+          iterations: ITERATIONS,
+          warmup: WARMUP,
+          operation,
+          options: isDeserialize(operation)
+            ? { ...OPTIONS[operation], promoteValues }
+            : OPTIONS[operation]
+        });
       }
     }
   }
