+++ test/bench/granular/regexp.bench.ts
@@ -1,0 +1,30 @@
+import { Suite } from 'bson-bench';
+import {
+  getTestDocs,
+  runSuiteAndWriteResults,
+  LIBRARY_SPEC,
+  OPERATIONS,
+  ITERATIONS,
+  WARMUP
+} from './common';
+
+async function main() {
+  const suite = new Suite('Regex');
+
+  const testDocs = await getTestDocs('regex');
+
+  for (const operation of OPERATIONS) {
+    for (const documentPath of testDocs) {
+      suite.task({
+        documentPath,
+        library: LIBRARY_SPEC,
+        iterations: ITERATIONS,
+        warmup: WARMUP,
+        operation,
+        options: {}
+      });
+    }
+  }
+  await runSuiteAndWriteResults(suite);
+}
+main();
