+++ test/bench/tsconfig.json
@@ -23,6 +23,6 @@
   },
   "include": [
     "./src/**/*.ts",
-    "./*.ts"
+    "./granular/*.ts"
   ]
 }
